// src/app/models/tenant-config.model.ts
export interface tenantConfigModel {
	id?: number;
	tenantId?: number;
	tipoNegocio?: string;
	since?: string;
	imgLogo?: string;
	history?: string;
	vision?: string;
	listVision?: string[];
	bussinesEmail?: string;
	twitter?: string;
	facebook?: string;
	linkedin?: string;
	instagram?: string;
	tiktok?: string;
	schedules?: string;
}